requirements.txt
aiogram~=3.0
environs~=9.0
redis
betterlogging

# # For enabling api:
backoff
ujson
yarl
aiohttp

# # For PostgreSQL sqlalchemy + alembic:
sqlalchemy~=2.0
alembic~=1.0
asyncpg

apscheduler
cachetools

aiogram_dialog
fluent.runtime~=0.4